+++ src/main/java/com/nexmo/client/redact/RedactRequest.java
@@ -1,0 +1,113 @@
+/*
+ * Copyright (c) 2011-2018 Nexmo Inc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.nexmo.client.redact;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonValue;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nexmo.client.NexmoUnexpectedException;
+
+/**
+ * Represents a request to the Redact API.
+ */
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class RedactRequest {
+
+    private String id;
+    private Product product;
+    private Type type;
+
+    /**
+     * Construct a RedactRequest object with all required fields.
+     *
+     * @param id      The transaction id to redact.
+     * @param product The {@link Product} that the id relates to.
+     */
+    public RedactRequest(String id, Product product) {
+        this.id = id;
+        this.product = product;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public Product getProduct() {
+        return product;
+    }
+
+    public void setProduct(Product product) {
+        this.product = product;
+    }
+
+    public Type getType() {
+        return type;
+    }
+
+    public void setType(Type type) {
+        this.type = type;
+    }
+
+    public String toJson() {
+        try {
+            ObjectMapper mapper = new ObjectMapper();
+            return mapper.writeValueAsString(this);
+        } catch (JsonProcessingException jpe) {
+            throw new NexmoUnexpectedException("Failed to produce json from RedactRequest object.", jpe);
+        }
+    }
+
+    public enum Product {
+        SMS("sms"),
+        VOICE("voice"),
+        NUMBER_INSIGHTS("number-insight"),
+        VERIFY("verify"),
+        VERIFY_SDK("verify-sdk"),
+        MESSAGE("message"),
+        WORKFLOW("workflow");
+
+        private String value;
+
+        Product(String value) {
+            this.value = value;
+        }
+
+        @JsonValue
+        public String getValue() {
+            return value;
+        }
+    }
+
+    public enum Type {
+        INBOUND, OUTBOUND;
+
+        @JsonValue
+        public String getValue() {
+            return this.name().toLowerCase();
+        }
+    }
+}